% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imis_api.R
\name{osha_search}
\alias{osha_search}
\title{Search of the OSHA IMIS system for internal tracking of citations & inspections}
\usage{
osha_search(
  std_query,
  sic = "",
  office = "All",
  p_logger = 1,
  p_start = 0,
  p_finish = 0,
  p_show = 50,
  start_month = stringr::str_pad(lubridate::month(Sys.Date() - 90), 2, "left", "0"),
  start_day = stringr::str_pad(lubridate::mday(Sys.Date() - 90), 2, "left", "0"),
  start_year = stringr::str_pad(lubridate::year(Sys.Date() - 90), 2, "left", "0"),
  end_month = stringr::str_pad(lubridate::month(Sys.Date()), 2, "left", "0"),
  end_day = stringr::str_pad(lubridate::mday(Sys.Date()), 2, "left", "0"),
  end_year = stringr::str_pad(lubridate::year(Sys.Date()), 2, "left", "0"),
  category = "",
  InspNr = ""
)
}
\arguments{
\item{std_query}{the search query}

\item{sic}{unsure}

\item{office}{seems to be a regional OSHA office that the observation was filed in}

\item{p_logger}{unsure}

\item{p_start}{pagination start row}

\item{p_finish}{pagination finish row}

\item{p_show}{how many results to show per page}

\item{start_month}{start of query month in "mm" format}

\item{start_day}{start of query day in "dd" format}

\item{start_year}{start of query year in "yyyy" format}

\item{end_month}{end of query month in "mm" format}

\item{end_day}{end of query day in "dd" format}

\item{end_year}{end of query year in "yyyy" format}

\item{category}{unsure}

\item{InspNr}{unsure}
}
\value{
returns a dataframe with date ,location, and details of citations and inspections.
}
\description{
Search of the OSHA IMIS system for internal tracking of citations & inspections
}
