% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imis_api.R
\name{osha_search}
\alias{osha_search}
\title{Search of the OSHA IMIS system for internal tracking of citations & inspections}
\usage{
osha_search(
  std_query,
  sic = "",
  office = "All",
  p_logger = 1,
  p_start = 0,
  p_finish = 0,
  p_show = 50,
  start_date = Sys.Date() - 365,
  end_date = Sys.Date(),
  category = "",
  InspNr = ""
)
}
\arguments{
\item{std_query}{the search query}

\item{sic}{unsure}

\item{office}{seems to be a regional OSHA office that the observation was filed in}

\item{p_logger}{unsure}

\item{p_start}{pagination start row}

\item{p_finish}{pagination finish row}

\item{p_show}{how many results to show per page}

\item{start_date}{"yyyymmdd" start date to limit results}

\item{end_date}{"yyyymmdd" end date to limit results. Defaults to last 365 days}

\item{category}{unsure}

\item{InspNr}{unsure}
}
\value{
returns a dataframe with date ,location, and details of citations and inspections.
}
\description{
Search of the OSHA IMIS system for internal tracking of citations & inspections
}
